requirements.txt
# Image Search 프로젝트 의존성

# 기본 라이브러리
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
datasets>=2.10.0
Pillow>=9.0.0
numpy>=1.21.0
scikit-learn>=1.0.0

# 임베딩 모델
sentence-transformers>=2.2.0

# 유틸리티
tqdm>=4.64.0
requests>=2.28.0

# 선택적 의존성 (성능 향상)
accelerate>=0.20.0
safetensors>=0.3.0
huggingface_hub>=0.16.0

# 개발 도구 (선택사항)
jupyter>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
